split up tensormeta
